Add reCAPTCHA to your website Add reCAPTCHA to your forms and actions you want to protect. The example below automatically renders the reCAPTCHA widget on an HTML element. Learn more about explicitly rendering widgets .
Load the JavaScript API.
On the element where the widget will render:
Add the g-recaptcha class.
Add the data-sitekey attribute and set it equal to your reCAPTCHA key.
Add the data-action attribute and set it to an action name. Learn more about actions .

<html>
  <head>
    <title>reCAPTCHA demo: Simple page</title>
    <script src="https://www.google.com/recaptcha/enterprise.js" async defer></script>
  </head>
  <body>
    <form action="" method="POST">
      <div class="g-recaptcha" data-sitekey="6LcSxFArAAAAAIYjPHUHkgQTx1ZP1EpxI-eRGoyO" data-action="LOGIN"></div>
      <br/>
      <input type="submit" value="Submit">
    </form>
  </body>
</html>

Send the response token to your application backend. The response token expires after two minutes.

_____
I want to implement this to signup action.